rewrote the tests according to the assignments
Random sequence.
Sequence of length 1.
Sequence where one element is repeated.
Sequence where all elements occur 1 time.
Sequence of length =0.
If an array is transmitted, try not to fill it completely.
If a formation is transmitted, try to transmit an array that does not start with 1.
If a formation is transmitted, try to transmit an array in which, for example, 2 occurs first before 3.
You can try an alphabet that does not correspond to the formation (more or less different elements than in the formation).
Alphabet can be without NullValue (should break).